Close popup on mouseUp if mouseDown is held
Another effort to replicate system behavior; Opening the popup by pressing the left mouse button and holding it allows the user to close the popup by lifting their finger off the mouse button. This applies even if the mouseUp is triggered above the popup to allow for quick actions. Take extra care to not dismiss the popup on an unrelated mouseUp event.